requirements.txt
Jinja2==3.1.3
markdown2==2.4.12
MarkupSafe==2.1.5
PyYAML==6.0.1
